feat: Add Inter font and update CSS styles
This commit adds the Inter font from Google Fonts and updates the CSS styles in the specified files to use the Inter font family. This improves the typography and enhances the overall visual appeal of the application.

The following files were modified:
- HProjects/src/app/[[...rest]]/index.css
- HProjects/src/components/modals/taskadd.jsx

The changes include importing the Inter font using a Google Fonts URL and applying the 'Inter' font family to the '.boxtexts' class in the CSS file.