Add 'added' field to logger mutation
This commit adds the 'added' field to the logger mutation in the 'updater.ts' file. The 'added' field is an optional boolean parameter that indicates whether a task has been added. When 'added' is true, the logger will record that a task has been created with the assigned user. This change improves the logging functionality for task creation.

Files modified:
- HProjects/convex/updater.ts