refactor: Remove team member from project
Remove a team member from the project by updating the otherusers array in the database. This change ensures that the user is no longer associated with the project and removes their access rights. The removed user will no longer appear in the team members list on the project settings page.